adding default storybook and types file
