+++ packages/react/ds/src/stack/stack.stories.tsx
@@ -76,6 +76,12 @@ export const Default: Story = {
   },
 };
 
+export const WithDefaultProps: Story = {
+  args: {
+    children: [<div>Item 1</div>, <div>Item 2</div>, <div>Item 3</div>],
+  },
+};
+
 export const ResponsiveDirectionWithDivider: Story = {
   args: {
     direction: { base: 'row', sm: 'row', md: 'column' },
