+++ packages/react/ds/src/stack/types.ts
@@ -1,0 +1,44 @@
+import { ReactNode } from 'react';
+
+export type Gap =
+  | number
+  | {
+      base?: number;
+      xs?: number;
+      sm?: number;
+      md?: number;
+      lg?: number;
+      xl?: number;
+      '2xl'?: number;
+    };
+type SimpleDirection = 'column' | 'row';
+
+export type Alignment = 'start' | 'center' | 'end';
+export type Distribution =
+  | 'start'
+  | 'center'
+  | 'end'
+  | 'between'
+  | 'around'
+  | 'evenly';
+export type Direction =
+  | SimpleDirection
+  | {
+      base?: SimpleDirection;
+      xs?: SimpleDirection;
+      sm?: SimpleDirection;
+      md?: SimpleDirection;
+      lg?: SimpleDirection;
+      xl?: SimpleDirection;
+      '2xl'?: SimpleDirection;
+    };
+export type StackProps = {
+  direction?: Direction;
+  itemsAlignment?: Alignment;
+  itemsDistribution?: Distribution;
+  gap?: Gap;
+  hasDivider?: boolean;
+  wrap?: boolean;
+  fixedHeight?: string;
+  children: ReactNode;
+};
